script attempt/azure_requirements.txt
azureml.core
ipykernel
matplotlib
numpy
pandas
scikit-learn==1.1.1
seaborn
xgboost